# Give optimisation hints to compiler