--- a/src/backend/looped/self/persist_outfit.cpp
+++ b/src/backend/looped/self/persist_outfit.cpp
@@ -15,13 +15,13 @@ namespace big
 		int offset = 0;
 		switch (model)
 		{
-		case "mp_m_freemode_01"_J: break;
-		case "mp_f_freemode_01"_J:
-		{
-			offset = 1;
-			break;
-		}
-		default: return false; //For non-normal models
+			case "mp_m_freemode_01"_J: [[likely]] break;
+			case "mp_f_freemode_01"_J: [[likely]]
+			{
+				offset = 1;
+				break;
+			}
+			default:[[unlikely]] return false; //For non-normal models
 		}
 
 		return PED::GET_PED_DRAWABLE_VARIATION(self::ped, ComponentId::AUXILIARY) == 15 && PED::GET_PED_DRAWABLE_VARIATION(self::ped, ComponentId::TORSO) == 15 && PED::GET_PED_DRAWABLE_VARIATION(self::ped, ComponentId::LEGS) == (14 + offset);
@@ -31,17 +31,17 @@ namespace big
 	{
 		//Disable clothing validation
 		*scr_globals::reset_clothing.as<PBOOL>() = FALSE;
-		if (auto tunable = g_tunables_service->get_tunable<PBOOL>("DISABLE_CLOTHING_SAVE_SLOT_VALIDATION"_J))
+		if (auto tunable = g_tunables_service->get_tunable<PBOOL>("DISABLE_CLOTHING_SAVE_SLOT_VALIDATION"_J)) [[likely]]
 			*tunable = TRUE;
 
 		if (g.self.persist_outfit.empty())
 			return; //Off
 
-		if (g_local_player == nullptr || PED::IS_PED_DEAD_OR_DYING(self::ped, TRUE) || STREAMING::IS_PLAYER_SWITCH_IN_PROGRESS() || DLC::GET_IS_LOADING_SCREEN_ACTIVE())
+		if (g_local_player == nullptr || PED::IS_PED_DEAD_OR_DYING(self::ped, TRUE) || STREAMING::IS_PLAYER_SWITCH_IN_PROGRESS() || DLC::GET_IS_LOADING_SCREEN_ACTIVE()) [[unlikely]]
 			return; //Dead or Loading
 
 		auto model = ENTITY::GET_ENTITY_MODEL(self::ped);
-		if (is_taking_shower(model))
+		if (is_taking_shower(model)) [[unlikely]]
 			return; //Showering
 
 		if (g.self.persist_outfits_mis && NETWORK::NETWORK_IS_ACTIVITY_SESSION())
@@ -55,10 +55,10 @@ namespace big
 			persisting_outfit                   = g.self.persist_outfit;
 			folder saved_outfit_path            = g_file_manager.get_project_folder("saved_outfits");
 			const auto persist_outfit_file_path = saved_outfit_path.get_file(persisting_outfit).get_path();
-			if (std::filesystem::exists(persist_outfit_file_path))
+			if (std::filesystem::exists(persist_outfit_file_path)) [[likely]]
 			{
 				std::ifstream i(persist_outfit_file_path);
-				if (i.is_open())
+				if (i.is_open()) [[likely]]
 				{
 					outfit.clear();
 					try