--- a/src/backend/looped/self/unlimited_oxygen.cpp
+++ b/src/backend/looped/self/unlimited_oxygen.cpp
@@ -11,7 +11,7 @@ namespace big
 
 		virtual void on_tick() override
 		{
-			if (g_local_player)
+			if (g_local_player) [[likely]]
 				g_local_player->m_ped_intelligence->m_oxygen_time = 0;
 		}
 	};