--- a/src/backend/looped/system/self_globals.cpp
+++ b/src/backend/looped/system/self_globals.cpp
@@ -9,9 +9,9 @@ namespace big
 	void looped::system_self_globals()
 	{
 		if (!*g_pointers->m_gta.m_network_player_mgr || !(*g_pointers->m_gta.m_network_player_mgr)->m_local_net_player
-		    || (*g_pointers->m_gta.m_network_player_mgr)->m_local_net_player->m_player_id == -1)
+		    || (*g_pointers->m_gta.m_network_player_mgr)->m_local_net_player->m_player_id == -1) [[unlikely]]
 			self::id = 0;
-		else
+		else [[likely]]
 			self::id = (*g_pointers->m_gta.m_network_player_mgr)->m_local_net_player->m_player_id;
 
 		self::ped = PLAYER::PLAYER_PED_ID();