--- a/src/hooks/fuzzer/read_bits_single.cpp
+++ b/src/hooks/fuzzer/read_bits_single.cpp
@@ -8,7 +8,7 @@ namespace big
 	{
 		auto orig = g_hooking->get_original<hooks::read_bits_single>()(data, out_value, size, offset);
 
-		if (fuzzer::is_fuzzer_enabled())
+		if (fuzzer::is_fuzzer_enabled()) [[unlikely]]
 		{
 			fuzzer::handle_vtable_write(out_value, sizeof(int), __FUNCTION__);
 			*out_value = fuzzer::fuzz_bits(*out_value, size);