--- a/src/hooks/fuzzer/sync_data_reader.cpp
+++ b/src/hooks/fuzzer/sync_data_reader.cpp
@@ -7,7 +7,7 @@ namespace big
 	{
 		auto orig = g_hooking->m_sync_data_reader_hook.get_original<decltype(&hooks::sync_reader_serialize_dword)>(1)(_this, dword, size);
 
-		if (fuzzer::is_fuzzer_enabled())
+		if (fuzzer::is_fuzzer_enabled()) [[unlikely]]
 		{
 			fuzzer::handle_vtable_write(dword, sizeof(uint32_t), __FUNCTION__);
 			*dword = fuzzer::fuzz_bits(*dword, size);
@@ -21,7 +21,7 @@ namespace big
 	{
 		auto orig = g_hooking->m_sync_data_reader_hook.get_original<decltype(&hooks::sync_reader_serialize_word)>(2)(_this, word, size);
 
-		if (fuzzer::is_fuzzer_enabled())
+		if (fuzzer::is_fuzzer_enabled()) [[unlikely]]
 		{
 			fuzzer::handle_vtable_write(word, sizeof(uint16_t), __FUNCTION__);
 			*word = fuzzer::fuzz_bits(*word, size);
@@ -35,7 +35,7 @@ namespace big
 	{
 		auto orig = g_hooking->m_sync_data_reader_hook.get_original<decltype(&hooks::sync_reader_serialize_byte)>(3)(_this, byte, size);
 
-		if (fuzzer::is_fuzzer_enabled())
+		if (fuzzer::is_fuzzer_enabled()) [[unlikely]]
 		{
 			fuzzer::handle_vtable_write(byte, sizeof(uint8_t), __FUNCTION__);
 			*byte = fuzzer::fuzz_bits(*byte, size);
@@ -49,7 +49,7 @@ namespace big
 	{
 		auto orig = g_hooking->m_sync_data_reader_hook.get_original<decltype(&hooks::sync_reader_serialize_int32)>(4)(_this, i, size);
 
-		if (fuzzer::is_fuzzer_enabled())
+		if (fuzzer::is_fuzzer_enabled()) [[unlikely]]
 		{
 			fuzzer::handle_vtable_write(i, sizeof(int32_t), __FUNCTION__);
 			*i = fuzzer::fuzz_signed_bits(*i, size);
@@ -63,7 +63,7 @@ namespace big
 	{
 		auto orig = g_hooking->m_sync_data_reader_hook.get_original<decltype(&hooks::sync_reader_serialize_int16)>(5)(_this, i, size);
 
-		if (fuzzer::is_fuzzer_enabled())
+		if (fuzzer::is_fuzzer_enabled()) [[unlikely]]
 		{
 			fuzzer::handle_vtable_write(i, sizeof(int16_t), __FUNCTION__);
 			*i = fuzzer::fuzz_signed_bits(*i, size);
@@ -77,7 +77,7 @@ namespace big
 	{
 		auto orig = g_hooking->m_sync_data_reader_hook.get_original<decltype(&hooks::sync_reader_serialize_signed_byte)>(6)(_this, i, size);
 
-		if (fuzzer::is_fuzzer_enabled())
+		if (fuzzer::is_fuzzer_enabled()) [[unlikely]]
 		{
 			fuzzer::handle_vtable_write(i, sizeof(int8_t), __FUNCTION__);
 			*i = fuzzer::fuzz_signed_bits(*i, size);
@@ -91,7 +91,7 @@ namespace big
 	{
 		auto orig = g_hooking->m_sync_data_reader_hook.get_original<decltype(&hooks::sync_reader_serialize_bool)>(7)(_this, flag, size);
 
-		if (fuzzer::is_fuzzer_enabled())
+		if (fuzzer::is_fuzzer_enabled()) [[unlikely]]
 		{
 			fuzzer::handle_vtable_write(flag, sizeof(bool), __FUNCTION__);
 			*flag = fuzzer::fuzz_bool(*flag);
@@ -105,7 +105,7 @@ namespace big
 	{
 		auto orig = g_hooking->m_sync_data_reader_hook.get_original<decltype(&hooks::sync_reader_serialize_signed_float)>(16)(_this, flt, divisor, size);
 
-		if (fuzzer::is_fuzzer_enabled())
+		if (fuzzer::is_fuzzer_enabled()) [[unlikely]]
 		{
 			fuzzer::handle_vtable_write(flt, sizeof(float), __FUNCTION__);
 			*flt = fuzzer::fuzz_float(*flt, size, divisor, true);
@@ -119,7 +119,7 @@ namespace big
 	{
 		auto orig = g_hooking->m_sync_data_reader_hook.get_original<decltype(&hooks::sync_reader_serialize_float)>(17)(_this, flt, divisor, size);
 
-		if (fuzzer::is_fuzzer_enabled())
+		if (fuzzer::is_fuzzer_enabled()) [[unlikely]]
 		{
 			fuzzer::handle_vtable_write(flt, sizeof(float), __FUNCTION__);
 			*flt = fuzzer::fuzz_float(*flt, size, divisor, false);
@@ -133,7 +133,7 @@ namespace big
 	{
 		auto orig = g_hooking->m_sync_data_reader_hook.get_original<decltype(&hooks::sync_reader_serialize_net_id)>(18)(_this, id);
 
-		if (fuzzer::is_fuzzer_enabled())
+		if (fuzzer::is_fuzzer_enabled()) [[unlikely]]
 		{
 			fuzzer::handle_vtable_write(id, sizeof(uint16_t), __FUNCTION__);
 			*id = fuzzer::fuzz_network_id(*id);
@@ -147,7 +147,7 @@ namespace big
 	{
 		auto orig = g_hooking->m_sync_data_reader_hook.get_original<decltype(&hooks::sync_reader_serialize_vec3)>(19)(_this, vec, divisor, size);
 
-		if (fuzzer::is_fuzzer_enabled())
+		if (fuzzer::is_fuzzer_enabled()) [[unlikely]]
 		{
 			fuzzer::handle_vtable_write(vec, sizeof(rage::fvector3), __FUNCTION__);
 			*vec = fuzzer::fuzz_vector(*vec, size, divisor, false);
@@ -161,7 +161,7 @@ namespace big
 	{
 		auto orig = g_hooking->m_sync_data_reader_hook.get_original<decltype(&hooks::sync_reader_serialize_vec3_signed)>(21)(_this, vec, divisor, size);
 
-		if (fuzzer::is_fuzzer_enabled())
+		if (fuzzer::is_fuzzer_enabled()) [[unlikely]]
 		{
 			fuzzer::handle_vtable_write(vec, sizeof(rage::fvector3), __FUNCTION__);
 			*vec = fuzzer::fuzz_vector(*vec, size, divisor, true);
@@ -175,7 +175,7 @@ namespace big
 	{
 		auto orig = g_hooking->m_sync_data_reader_hook.get_original<decltype(&hooks::sync_reader_serialize_array)>(23)(_this, array, size);
 
-		if (fuzzer::is_fuzzer_enabled())
+		if (fuzzer::is_fuzzer_enabled()) [[unlikely]]
 		{
 			fuzzer::handle_vtable_write(array, size / 8, __FUNCTION__);
 			fuzzer::fuzz_data(array, size / 8);