--- a/src/hooks/gui/wndproc.cpp
+++ b/src/hooks/gui/wndproc.cpp
@@ -6,7 +6,7 @@ namespace big
 {
 	LRESULT hooks::wndproc(HWND hwnd, UINT msg, WPARAM wparam, LPARAM lparam)
 	{
-		if (g_running)
+		if (g_running) [[likely]]
 		{
 			g_renderer.wndproc(hwnd, msg, wparam, lparam);
 		}