--- a/src/hooks/misc/http_start_request.cpp
+++ b/src/hooks/misc/http_start_request.cpp
@@ -7,7 +7,7 @@ namespace big
 {
     bool hooks::http_start_request(void* request, const char* uri)
     {
-        if (strstr(uri, "Bonus"))
+        if (strstr(uri, "Bonus")) [[unlikely]]
         {
             // This is for worst case scenario where a report does slip through the cracks...
             // Lets make it go somewhere it doesn't matter -- don't let the reports reach their servers!