--- a/src/hooks/misc/queue_dependency.cpp
+++ b/src/hooks/misc/queue_dependency.cpp
@@ -7,11 +7,11 @@ namespace big
 {
 	bool inline is_address_in_game_region(int64_t address)
 	{
-		if(!address)
+		if(!address) [[unlikely]]
 			return false;
 		static int64_t moduleBase = NULL;
 		static int64_t moduleSize = NULL;
-		if (!moduleBase || !moduleSize)
+		if (!moduleBase || !moduleSize) [[unlikely]]
 		{
 			MODULEINFO info;
 			if (!GetModuleInformation(GetCurrentProcess(), GetModuleHandle(0), &info, sizeof(info)))
@@ -42,10 +42,10 @@ namespace big
 		int64_t f2 = *reinterpret_cast<int64_t*>(cb + 0x100);
 		int64_t f3 = *reinterpret_cast<int64_t*>(cb + 0x1A0);
 
-		if (!is_address_in_game_region(f1) || !is_address_in_game_region(f2) || !is_address_in_game_region(f3))
+		if (!is_address_in_game_region(f1) || !is_address_in_game_region(f2) || !is_address_in_game_region(f3)) [[likely]]
 			return false;
 
-		if(*reinterpret_cast<uint8_t*>(f1) != 0xE9)
+		if(*reinterpret_cast<uint8_t*>(f1) != 0xE9) [[likely]]
 			return false;
 
 		return true;