--- a/src/hooks/protections/fragment_physics_crash.cpp
+++ b/src/hooks/protections/fragment_physics_crash.cpp
@@ -4,7 +4,7 @@ namespace big
 {
 	bool hooks::fragment_physics_crash_2(float* a1, float* a2)
 	{
-		if (!a1 || !a2)
+		if (!a1 || !a2) [[unlikely]]
 			return false;
 
 		return g_hooking->get_original<hooks::fragment_physics_crash_2>()(a1, a2);