--- a/src/hooks/protections/handle_join_request.cpp
+++ b/src/hooks/protections/handle_join_request.cpp
@@ -11,7 +11,7 @@ namespace big
 	bool hooks::handle_join_request(Network* network, rage::snSession* session, rage::rlGamerInfo* player_info, CJoinRequestContext* ctx, BOOL is_transition_session)
 	{
 		if (auto player = g_player_database_service->get_player_by_rockstar_id(player_info->m_gamer_handle.m_rockstar_id);
-		    player && player->block_join)
+		    player && player->block_join) [[unlikely]]
 		{
 			CMsgJoinResponse response{};
 			response.m_status_code = player->block_join_reason;
@@ -20,7 +20,7 @@ namespace big
 			    std::vformat("BLOCK_JOIN_INFO"_T, std::make_format_args(player->name)));
 			return false;
 		}
-		else
+		else [[likely]]
 		{
 			return g_hooking->get_original<hooks::handle_join_request>()(network, session, player_info, ctx, is_transition_session);
 		}