… src/hooks/protections/serialize_parachute_task.cpp
… src/hooks/protections/serialize_take_off_ped_variation_task.cpp
… src/hooks/protections/task_ambient_clips.cpp
… src/hooks/protections/task_parachute_object.cpp
… src/hooks/script/run_script_threads.cpp
… src/invoker/invoker.hpp
… src/native_hooks/all_scripts.hpp
… src/renderer/renderer.cpp
… src/services/persist_weapons/persist_weapons.cpp
… src/thread_pool.cpp