I split up my work into multiple commits as I did NO TESTING whatsoever to see if they make a difference, they can be reverted at any point in time.

This should definitely not be forced on every branch that exists in YimMenu, hot code paths should be prioritized (game thread mainly and expensive operations).