@xiaoxiao921 2024-04-09
Bad feeling about this, could this be atleast profiled with atleast a stopwatch or something? feel like this could backfire