@Yimura 2024-04-09
> Bad feeling about this, could this be atleast profiled with atleast a stopwatch or something? feel like this could backfire

It can definitely backfire and that's why it shouldn't be overused or forced into YimMenu. However it is a very powerful tool and can be used to improve certain things that are really inefficient.