@gir489returns 2024-04-09
> Bad feeling about this, could this be atleast profiled with atleast a stopwatch or something? feel like this could backfire

Just like with inline, ultimately the decision is still up to the compiler, but if you look here: https://en.cppreference.com/w/cpp/language/attributes/likely you can see at the bottom they produced a test case scenario that was extremely close to std::cos. std::cos is probably extremely optimized, but because of the likely/unlikely statements, you can tell the compiler how to generate the branch prediction code in such a way that the happy path scenario is usually the predicted path.