@gir489returns 2024-04-09
So I have come up with the following guidelines to use the likely/unlikely scenarios.

1. They should only be applied to scenarios where the function runs an indeterminate number of times (hooks/frame/render time dependent, all ad-hock function calls be left to the compiler for determination)
2. Any if statements that rely on user-configuration never be given a likely/unlikely scenario as we can never be predeterministic as to what the client wants.
2b. The only exception to 2 is debug scenarios, debug configurations should never be run by the client and we should never expect them to, they are there only for developers who know what they're doing.
3. Do not optimize anything that requires specific control flow. (TBD)