@gir489returns 2024-04-09
This should conclude the majority of my optimization emplacements.