@Yimura 2024-04-09
We need to do some profiling because my system is showing negligible performance changes however that doesn't mean there is a difference for system that are less performant.