@gir489returns 2024-04-22
I just benchmarked this with queue_dependency.

With the optimizations:

`[16:21:34.5675681][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x318D08
[16:21:34.5675723][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 4us 600ns
[16:21:34.5720734][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x961FBC
[16:21:34.5720829][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 11us 100ns
[16:21:34.5720995][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0xA66114
[16:21:34.5721008][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 1us 400ns
[16:21:34.5726901][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x66F09C
[16:21:34.5726958][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 6us 700ns
[16:21:34.5727074][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x66F0C4
[16:21:34.5727093][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 2us 0ns
[16:21:34.5727208][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x66F0EC
[16:21:34.5727218][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 1us 100ns
[16:21:34.5727234][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x66F074
[16:21:34.5727245][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 1us 100ns
[16:21:34.5753427][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x335660
[16:21:34.5753459][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 3us 500ns
[16:21:34.5753517][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x335638
[16:21:34.5753563][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 4us 700ns
[16:21:34.5753847][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x4015B8
[16:21:34.5753857][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 1us 100ns
[16:21:34.5788160][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x45B164
[16:21:34.5788226][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 7us 600ns
[16:21:34.5788842][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x45A3A8
[16:21:34.5788856][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 1us 500ns
[16:21:34.5788875][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x5CAA60
[16:21:34.5788882][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 900ns`

Without:

`[16:28:21.6586132][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x961FBC
[16:28:21.6586299][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 18us 700ns
[16:28:21.6586454][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0xA66114
[16:28:21.6586472][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 1us 900ns
[16:29:14.6359202][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x45A3A8
[16:29:14.6359234][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 3us 400ns
[16:29:14.6359251][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x5CAA60
[16:29:14.6359259][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 800ns
[16:29:14.6371187][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x318D08
[16:29:14.6371224][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 4us 200ns
[16:29:14.6386673][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x66F09C
[16:29:14.6386738][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 7us 700ns
[16:29:14.6386797][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x66F0C4
[16:29:14.6386813][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 1us 800ns
[16:29:14.6386924][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x66F0EC
[16:29:14.6386940][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 1us 800ns
[16:29:14.6386971][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x66F074
[16:29:14.6386986][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 1us 600ns
[16:29:14.6410049][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x335660
[16:29:14.6410080][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 3us 300ns
[16:29:14.6410139][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x335638
[16:29:14.6410156][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 1us 800ns
[16:29:14.6410397][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x4015B8
[16:29:14.6410415][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 1us 900ns
[16:29:14.6441614][INFO/queue_dependency.cpp:65] Blocking AC Verifier 0x45B164
[16:29:14.6441681][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 7us 700ns`

Random calls with:

`[16:21:34.5791955][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5791966][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5791976][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 0ns
[16:21:34.5791986][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792010][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792028][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792038][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792047][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 0ns
[16:21:34.5792059][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792071][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792085][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 400ns
[16:21:34.5792097][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792107][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792134][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792153][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792167][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792181][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792191][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 0ns
[16:21:34.5792217][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792236][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792248][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792264][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792275][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792303][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:21:34.5792322][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns`

Random calls without:

`[16:29:14.6443151][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 600ns
[16:29:14.6443213][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 200ns
[16:29:14.6443393][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 200ns
[16:29:14.6443768][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 400ns
[16:29:14.6443840][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:29:14.6443864][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:29:14.6443882][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:29:14.6443895][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 0ns
[16:29:14.6443910][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:29:14.6443943][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:29:14.6443967][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 100ns
[16:29:14.6443979][INFO/benchmark.hpp:18] queue_dependency finished with a resulting time of: 0us 0ns`

I think it's safe to say that the branch prediction influence has an immense difference and should be considered ready for merging.